Enable dependabot for actions and python
Enable dependabot for the GHA and Python ecosystem.
The GHA one lets us keep actions in workflows up to date.
The Python one updates all dependencies (including extras) defined
in pyproject.toml.

A new workflow is used to add a commit on top of dependabot's
Python updates. The new commit includes the requirements updates
generated by recompiling pyproject.toml. This is required so
that CI may pass.

Dependanbot Pytorch ecosystem is configured to bundle updates as
much as possible so to avoid too many patches and merge conflicts.
That can be switched back to individual PRs at any time if we
prefer to update packages individually.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>