Use spyre-bot GitHub app token
In the dependabot workflow, that updates dependabot PRs with
uv.lock and requirement updates, start using a token from
the Spyre-bot GitHub app.

This is required because commits created with GITHUB_TOKEN
do not trigger workflows (to avoid potential loops).

This PR also updates the check for changed files to
verify `uv.lock`, since the lock file was introduced.

Fixes: #135

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>